Move toolbar to menu activity
