+++ mobile/app/src/main/res/layout/activity_menu.xml
@@ -7,6 +7,14 @@
     android:gravity="center"
     android:padding="20dp">
 
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/menu_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        android:title="@string/app_name"
+        android:titleTextColor="@color/colorWhite" />
+
     <TextView
         android:id="@+id/nicknameLabel"
         android:layout_width="wrap_content"
@@ -75,29 +83,6 @@
         android:textColor="@color/colorWhite"
         android:textStyle="bold" />
 
-    <Button
-        android:id="@+id/Settings"
-        android:layout_width="200dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="10dp"
-        android:background="@color/colorGreen"
-        android:onClick="OpenSettings"
-        android:text="@string/menu_Settings"
-        android:textAllCaps="false"
-        android:textColor="@color/colorWhite"
-        android:textStyle="bold" />
 
-    <!-- New Account button -->
-    <Button
-        android:id="@+id/Account"
-        android:layout_width="200dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="10dp"
-        android:background="@color/colorGreen"
-        android:onClick="OpenAccount"
-        android:text="@string/login"
-        android:textAllCaps="false"
-        android:textColor="@color/colorWhite"
-        android:textStyle="bold" />
 
 </LinearLayout>
